agent/requirements.txt
openai>=1.60.0
langgraph>=0.2.46
pydantic>=2.8.0
pyyaml>=6.0.1
gitpython>=3.1.43
requests>=2.32.3
python-dotenv>=1.0.1
rich>=13.8.0
